break the circular dependency between ResourcePath and AutoConfigClass
